Ajout de debug pour essayer de comprendre ce qu'il se passe Modification de init du communicator pour inclure la definition de la fonction de read

+++ Robot/Software/ControlleurPrincipal.cpp
@@ -260,7 +260,7 @@ float ControlleurPrincipal::getAngleResultant(float depart, float angle, bool ga
 
     if (gauche)
     {
-		resultat = fmod(depart + angle, 360);
+		resultat = (float) fmod(depart + angle, 360);
     }
     else
     {
